feat: specified userId dashboard has to render
This allows a common user to only see its own dashboard
and admin users to see other user's dashboard.

When a user_id param is found in the URL, one will use this one
when calling the API.
When none is found, it will use the user_id cookie.

When the API calls are done, a verification is done backend to confirm
a user has the correct permissions to see someone else's dashboard,
via the JWT claims sent when doing the requests (httponly, invisible to javascript
but sent anyway).