feat: Added basic Achievement + Relation + Condition
This is a PoC about a new Achievement and its Relation + Conditions.
Only "Action Trigger" condition type is supported for now,
"required Achievement" and "Time Frame" will follow shortly.

Rewards + Assets will follow later.

Only a couple of Conditions will be implemented for Achievements,
the other ones are actually Actions conditions, missing for now.
These are conditions needed for an Action to be 'receivable'.