fix: only GameDetails must set $game store to a valid value
Setting $game.set(null) in Settings component (when destroying a game)
was incorrect because it was made before gameUrl was set to null in Dashboard,
leading to an incorrect state: GameDetails and children were trying to render
with a null $game store, while they shouldn't be instantiated by Dashboard.

This was still the case, gameUrl not being set to null yet.